split/refactor test_sync_utils.py.test_config_filtering between test_config.py and test_sync_utils.py
